requirements.txt
openai-agents
python-dotenv 
gitingest
fastapi
uvicorn[standard]
jinja2
python-multipart
api-analytics[fastapi]
langgraph
pydantic>=2.0.0
pydantic-settings>=2.0.0